Refactor strip and index.html for easier updates
Color settings are now automatically pulled into the index.html and
returned from the `/settings` endpoint. This makes it easier to add
new colors in the future and have them automatically populated in the
necessary places.

To add a new color, only the strip's settings hash and function need to
be created, then it can be used from the chatbot and index.html page.